% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{deamidation_coeffs}
\alias{deamidation_coeffs}
\title{Calculate the coefficients of each of the non or deamidated peptides
for a given global q2e}
\usage{
deamidation_coeffs(q2e, max_Q, ndeam)
}
\arguments{
\item{max_Q}{}
}
\description{
Calculate the coefficients of each of the non or deamidated peptides
for a given global q2e
}
