% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{design}
\alias{design}
\title{Title}
\usage{
design(lgth_g, p, isotopes)
}
\arguments{
\item{lgth_g}{number of parameters}

\item{p}{peptide index}

\item{isotopes}{theoretical isotopic distributions}
}
\value{
Design matrix
}
\description{
Title
}
