% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{extractPeaks}
\alias{extractPeaks}
\title{Title}
\usage{
extractPeaks(x, ...)
}
\arguments{
\item{...}{}
}
\value{
data.frame with mass, intensity and s2n values for pseudo-isotopic
clusters
}
\description{
Title
}
