% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_pqi.R
\name{fvsr}
\alias{fvsr}
\title{Fitted versus residuals plot per peptide}
\usage{
fvsr(pqi_m, title = "")
}
\arguments{
\item{pqi_m}{data.frame of model estimates per replicate and peptide}

\item{title}{}
}
\description{
Fitted versus residuals plot per peptide
}
