% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{get_isodists}
\alias{get_isodists}
\title{Get isotopic distributions}
\usage{
get_isodists(seqs, nhyds)
}
\arguments{
\item{nhyds}{}
}
\value{
Array of intensity distributions, without the masses
}
\description{
Wrapper for ballite ms_iso.
Calculates theoretical isotpic deistributions for a list of
sequences
}
