% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{get_ref_isopeaks}
\alias{get_ref_isopeaks}
\title{Get isotopic peaks from reference dataset}
\usage{
get_ref_isopeaks(which_params = F, params = NULL)
}
\arguments{
\item{params}{list of parameters and values for which the pre-computed iso_peaks is returned
Parameters supported: SNR, iterations, hws_smooth, halfWindowSize}

\item{list_params}{logical. If TRUE returns a dataframe of possible parameters
for which iso peaks has been precomputed for the reference dataset}
}
\value{
A list of dataframes, 1 per sample. Each dataframe has 3 columns,
m/z, intensity and signal-to-noise ratio for each of the n_isopeaks from each
peptide. Missing peaks are NAs.
}
\description{
Get isotopic peaks from reference dataset
}
\details{
It retrieves the isotopic envelopes for the default peptides for the
Orval Abbey dataset
}
\references{
Bethencourt, J.H. et al. (2022) ‘Data from “A biocodicological analysis of the medieval library and archive from Orval abbey, Belgium”’, Journal of open archaeology data, 10(0). doi:10.5334/joad.89.
}
