% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{getIsoPeaks}
\alias{getIsoPeaks}
\title{Preprocessing spectra for q2e estimation}
\usage{
getIsoPeaks(
  indir,
  readf = c("fread", "table", "mzml"),
  sep = "\\t",
  peptides_user = NULL,
  spectra = NULL,
  outdir = NULL,
  nchunks = 50,
  smooth_method = c("SavitzkyGolay", "Wavelet"),
  thresh.scale = 2.5,
  hws_smooth = 20,
  iterations = 20,
  halfWindowSize = 20,
  SNR = 0,
  tol = 0.00015,
  n_isopeaks = 6,
  min_isopeaks = 4,
  ncores = NULL,
  iocores = 1,
  vch = 10
)
}
\arguments{
\item{indir}{Folder containing spectra.}

\item{readf}{A string value. Choose function used to read spectra.
Currently restricted to one of "fread", "table" or "mzml".
See \code{\link[https://github.com/ismaRP/MALDIzooMS]{preprocessData}} for more info.}

\item{sep}{Separator character for txt spectra files}

\item{peptides_user}{A dataframe with peptide information. It must contain at least 3 columns,
peptide number or ID, name, and m/z. IF NULL, default are used, see details.
The number or ID must have the form Pep# and be in the first column.}

\item{outdir}{Folder where one peaks file for each sample should be written.
If NULL (default), no files are written.}

\item{nchunks}{Number of chunks all the spectra should be divided into for reading and
processing. If all spectra is loaded and processed at once, i.e. chunks=1,
it can overload RAM memory. If chunks>1 data is loaded and processed to the
much lighter list of peaks in chunks batches.}

\item{smooth_method}{Smootherning method, one of "SavitzkyGolay" or "Wavelet".
SavitzkyGolay uses \code{\link[MALDIquant]{smoothIntensity}} together with
\code{hws_smooth} parameter
Wavelet uses \code{\link[MALDIrppa]{wavSmoothing}} together with
\code{thresh.scale}}

\item{thresh.scale}{Smoothing factor for wavelet-based smoothing. Passed to \code{\link[MALDIrppa]{wavSmoothing}}}

\item{hws_smooth}{Half-window size parameter for smoothening. Passed to \code{\link[MALDIquant]{smoothIntensity}}}

\item{iterations}{Iterations parameter for baseline detection. Passed to \code{\link[MALDIquant]{estimateBaseline}}}

\item{halfWindowSize}{Half-window size parameter for local maximum detection. Passed to \code{\link[MALDIquant]{detectPeaks}}}

\item{SNR}{Signal to noise threshold for peak detection. Passed to \code{\link[MALDIquant]{detectPeaks}}}

\item{tol}{Tolerance factor for matching detected peaks to theoretical isotopic distribution.
Value used is m/z of monoisotopic peak * \code{tol}.}

\item{n_isopeaks}{Number of isotopic peaks to pick. Default is 5 and the maximum permitted.}

\item{min_isopeaks}{If less than min_isopeaks consecutive (about 1 Da difference) isotopic peaks
are detected, the whole isotopic envelope is discarded. Default is 4}

\item{ncores}{Number of cores used for the preprocessing of spectra. The cores will work in
parallel with the different spectra within a chunk.}

\item{iocores}{Number of cores used for I/O operations. For some systems I/O operations involving
multiple cores reading or writing at the same time from disk increases time.}
}
\value{
A list of dataframes, 1 per sample. Each dataframe has 3 columns,
m/z, intensity and signal-to-noise ratio for each of the n_isopeaks from each
peptide. Missing peaks are NAs.
}
\description{
Performs smoothening, baseline removal and peak detection on MALDI samples.
From the peaks, isotopic peaks for a list of peptides are extracted.
}
\details{
The default peptides are the ones from Nair et al. (2022).
The paper contains the details on the preprocessing procedure.
}
\references{
Nair, B. et al. (2022) ‘Parchment Glutamine Index (PQI): A novel method to estimate glutamine deamidation levels in parchment collagen obtained from low-quality MALDI-TOF data’, bioRxiv. doi:10.1101/2022.03.13.483627.
}
