% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{isotopic_deam}
\alias{isotopic_deam}
\title{Calculate the isotopic envelope of a deamidated peptide}
\usage{
isotopic_deam(q2e = 0.5, norm_func, max_Q, ...)
}
\arguments{
\item{q2e}{Input}

\item{max_Q}{Number of deamidating Q sites}

\item{...}{Relative intensity values of the isotopologues of the non-deamidated
versions of the peptide and the one with 1, 2,... deamidations}
}
\description{
Calculate the isotopic envelope of a deamidated peptide
}
