% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{isotopic_deam_df}
\alias{isotopic_deam_df}
\title{Isotopic distribution of deamidated peptdes
It produces the isotopic distirbution of peptides with a given
extent of deamidation}
\usage{
isotopic_deam_df(iso_peps, q2e = 0.5, norm_func = NULL, ...)
}
\arguments{
\item{q2e}{}
}
\description{
Isotopic distribution of deamidated peptdes
It produces the isotopic distirbution of peptides with a given
extent of deamidation
}
