% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{isotopic_deam_mat}
\alias{isotopic_deam_mat}
\title{Isotopic distribution of deamidated peptides
It produces the isotopic distribution of  peptides with a given extent
of deamidation.}
\usage{
isotopic_deam_mat(iso_peps, max_Q, q2e = 0.5, norm_func = NULL)
}
\arguments{
\item{q2e}{}
}
\description{
Isotopic distribution of deamidated peptides
It produces the isotopic distribution of  peptides with a given extent
of deamidation.
}
