% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wls_q2e.R
\name{lm_q2e}
\alias{lm_q2e}
\title{Fit a linear model without weights and forcing a 0 intercept}
\usage{
lm_q2e(
  norm_int = NULL,
  deam_0 = NULL,
  deam_1 = NULL,
  deam_2 = NULL,
  data = NULL,
  return_model = FALSE
)
}
\arguments{
\item{deam_2}{}
}
\description{
For use after group_by on samples and peptide index
}
\examples{
q2e_vals = peaks \%>\% filter(n_peaks > 0) \%>\%
   group_by(sample, replicate, pep_number) \%>\%
   summarise(lm_q2e(norm_int, weight, deam_0, deam_1, deam_2))
}
