% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{normalize_intensity}
\alias{normalize_intensity}
\title{Normalize vector of intensities by the maximum}
\usage{
normalize_intensity(intensity)
}
\arguments{
\item{intensity}{Vector of intensities}
}
\description{
Normalize vector of intensities by the maximum
}
