% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{peptide_labeller}
\alias{peptide_labeller}
\title{Labeler helper for plotting peptide spectra
Title}
\usage{
peptide_labeller(values)
}
\arguments{
\item{values}{}
}
\description{
Labeler helper for plotting peptide spectra
Title
}
