% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_plots.R
\name{plot_pept_spectra}
\alias{plot_pept_spectra}
\title{Title}
\usage{
plot_pept_spectra(
  indir,
  spectra_names,
  readf,
  peptides = NULL,
  n_isopeaks = 6,
  min_isopeaks = 4,
  tol = 0.00015,
  iter = 20,
  hws_smooth = 20,
  hws_peak = 20,
  snr = 0,
  ncores = NULL
)
}
\arguments{
\item{ncores}{}
}
\value{

}
\description{
Title
}
