% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wls_q2e.R
\name{plot_q}
\alias{plot_q}
\title{Plot distribution of q2e values per peptide}
\usage{
plot_q(q2e, peptides_user = NULL, label_idx = 2, label_func = label_value)
}
\arguments{
\item{q2e}{data.frame with q2e estimations per sample, replicate and peptide}

\item{peptides_user}{A dataframe with peptide information. It must contain at least 3 columns,
peptide number or ID, name, and m/z. If NULL, default are used, see\code{\link[MALDIutils]{getIsoPeaks}} details.
The number or ID must have the form Pep# and be in the first column.}

\item{label_idx}{Column index in peptides where the label is stored}

\item{label_func}{labeller function to process labels. See \code{\link[ggplot2]{labeller}}
Default is label_value.}
}
\description{
Plot distribution of q2e values per peptide
}
