% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{predict_sample}
\alias{predict_sample}
\title{Title}
\usage{
predict_sample(sample, replicate, pep_number, reliability, resp, pars)
}
\arguments{
\item{resp}{Response variable, either q or Logq}

\item{pars}{Contains the following parameters:
\describe{
\item{data_I_s}{part of the data frame corresponding to Sample s.}
\item{alpha}{estimate for fixed effect}
\item{sigma2_S}{estimate for variance for random effect of sample.}
\item{sigma2_R}{estimate for variance for random effect of replication.}
\item{gamma}{estimate for half power on Reliability}
\item{sigma2}{estimate for variance for on the different peptides. This should be a named vector.}
}}
}
\value{
tibble with predicted sample PQI and standard deviation
}
\description{
Title
}
