% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wls_q2e.R
\name{prepare_data}
\alias{prepare_data}
\title{Title}
\usage{
prepare_data(
  peptides,
  outdir = NULL,
  data_list = NULL,
  indir = NULL,
  n_isopeaks = 5
)
}
\arguments{
\item{n_isopeaks}{}
}
\value{

}
\description{
Title
}
