% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{prepare_peaks}
\alias{prepare_peaks}
\title{Prepare list of peaks data into a data.frame}
\usage{
prepare_peaks(
  peaks,
  peptides_user,
  n_isopeaks,
  int_column = "intensity",
  q2e = NULL
)
}
\arguments{
\item{peaks}{List of peaks matrix. Names are used as spectra name}

\item{peptides_user}{A dataframe with peptide information. It must contain at least 3 columns,
\code{pep_number} or ID, \code{mass}, \code{sequence} and \code{h_hyp} (# of hydroxyprolines).
IF NULL, default are used, see details.}

\item{n_isopeaks}{Number of isotopic peaks}

\item{q2e}{Optional q2e value to calculate the theoretical distribution of
a peptide with this extent of deamidation}
}
\value{
A data.frame with isotopic peaks detected from data, theoretical isotopic
envelopes for 1 and 2 deamidations and other associated data.
}
\description{
Prepare list of peaks data into a data.frame
}
\details{
The default peptides are the ones from Nair et al. (2022).
The paper contains the details on the preprocessing procedure.
}
