% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_data}
\alias{preprocess_data}
\title{preprocess_data}
\usage{
preprocess_data(
  indir,
  readf,
  peptides = NULL,
  outdir = NULL,
  chunks = 50,
  n_isopeaks = 6,
  min_isopeaks = 4,
  snr = 0,
  tol = 0.00015,
  iter = 20,
  hws_smooth = 20,
  hws_peak = 20,
  ncores = NULL
)
}
\arguments{
\item{ncores}{}
}
\value{
A list of mass-intensity-s2n peak
}
\description{
preprocess_data
}
