% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_pqi.R
\name{pept_qqplot}
\alias{pept_qqplot}
\title{Quantile-Quantile plots of the linear mixed effect estimates per peptide}
\usage{
pept_qqplot(pqi_m, title = "")
}
\arguments{
\item{pqi_m}{data.frame of model estimates per replicate and peptide}

\item{title}{Plot title}
}
\description{
Quantile-Quantile plots of the linear mixed effect estimates per peptide
}
