% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_plots.R
\name{plot_pept_spectra}
\alias{plot_pept_spectra}
\title{Title}
\usage{
plot_pept_spectra(
  indir,
  spectra_names,
  readf,
  peptides_user = NULL,
  label_idx = 2,
  label_func = label_value,
  normalize = F,
  smooth_method = c("SavitzkyGolay", "Wavelet"),
  thresh.scale = 2.5,
  hws_smooth = 8,
  iterations = 20,
  halfWindowSize = 20,
  SNR = 0,
  n_isopeaks = 5,
  min_isopeaks = 4,
  tol = 0.00015
)
}
\arguments{
\item{indir}{Folder containing spectra.}

\item{spectra_names}{Character vector. Spectra names to plot}

\item{readf}{A string value. choose function to use to read spectra.
Currently restricted to one of "fread", "table" or "mzml"}

\item{peptides_user}{A dataframe with peptide information. It must contain at least 3 columns,
peptide number or ID, name, and m/z. If NULL, default are used, see \code{\link[https://github.com/ismaRP/MALDIzooMS]{getIsoPeaks}} details.
The number or ID must have the form Pep# and be in the first column.}

\item{label_idx}{Column index in peptides where the label is stored}

\item{label_func}{labeller function to process labels. See \code{\link[ggplot2]{labeller}}
Default is label_value.}

\item{normalize}{Logical. Whether intensity hsould be normalized divinding by
the total intensity for each peptide peaks range. DEfault is False}

\item{smooth_method}{Smootherning method, one of "SavitzkyGolay" or "Wavelet".
SavitzkyGolay uses \code{\link[MALDIquant]{smoothIntensity}} together with
\code{hws_smooth} parameter
Wavelet uses \code{\link[MALDIrppa]{wavSmoothing}} together with
\code{thresh.scale}}

\item{thresh.scale}{Smoothing factor for wavelet-based smoothing. Passed to \code{\link[MALDIrppa]{wavSmoothing}}}

\item{hws_smooth}{Half-window size parameter for smoothening. Passed to \code{\link[MALDIquant]{smoothIntensity}}}

\item{iterations}{Iterations parameter for baseline detection. Passed to \code{\link[MALDIquant]{estimateBaseline}}}

\item{halfWindowSize}{Half-window size parameter for local maximum detection. Passed to \code{\link[MALDIquant]{detectPeaks}}}

\item{SNR}{Signal to noise threshold for peak detection. Passed to \code{\link[MALDIquant]{detectPeaks}}}

\item{n_isopeaks}{Number of isotopic peaks to pick. Default is 5 and the maximum permitted.}

\item{min_isopeaks}{If less than min_isopeaks consecutive (about 1 Da difference) isotopic peaks
are detected, the whole isotopic envelope is discarded. Default is 4}

\item{tol}{Tolerance factor for matching detected peaks to theoretical isotopic distribution.
Value used is m/z of monoisotopic peak * \code{tol}.}
}
\value{
ggplot with raw nd preprocessed spectra
}
\description{
Title
}
