% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectra_plots.R
\name{plot_preprocessing}
\alias{plot_preprocessing}
\title{Title}
\usage{
plot_preprocessing(
  m,
  p_name,
  s,
  s_br,
  p,
  baseline,
  noise,
  s_name,
  n_isopeaks,
  min_isopeaks,
  tol,
  normalize
)
}
\arguments{
\item{normalize}{}
}
\description{
Title
}
