% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_pqi.R
\name{plot_q}
\alias{plot_q}
\title{Plot distribution of q2e values per peptide}
\usage{
plot_q(q2e, peptides)
}
\arguments{
\item{q2e}{data.frame with q2e estimations per sample, replicate and peptide}

\item{peptides}{A dataframe with peptide information. It must contain at least 3 columns,
peptide number or ID, name, and m/z. IF NULL, default are used, see details.}
}
\description{
Plot distribution of q2e values per peptide
}
