% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lme_pqi.R
\name{predict_pqi}
\alias{predict_pqi}
\title{Title}
\usage{
predict_pqi(q2e, estimates, model, logq = T)
}
\arguments{
\item{q2e}{q2e table for each sample, replicate and peptide. q is stored in
a column called "resp"}

\item{estimates}{model estimates}

\item{model}{lme model object}

\item{logq}{whether q is in log-scale in q2e}
}
\description{
Title
}
