% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wls_q2e.R
\name{prepare_data}
\alias{prepare_data}
\title{Helper function to prepare data for q2e estimation}
\usage{
prepare_data(
  peptides_user = NULL,
  outdir = NULL,
  data_list = NULL,
  indir = NULL,
  n_isopeaks = 5
)
}
\arguments{
\item{n_isopeaks}{}
}
\description{
Helper function to prepare data for q2e estimation
}
