% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_spectra}
\alias{preprocess_spectra}
\title{Title}
\usage{
preprocess_spectra(s, ...)
}
\arguments{
\item{...}{}
}
\value{
peaks list, with mass, intensity and s2n
}
\description{
Title
}
