% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{preprocess_spectra}
\alias{preprocess_spectra}
\title{preprocess_spectra}
\usage{
preprocess_spectra(
  s,
  smoothf,
  iterations,
  halfWindowSize,
  SNR,
  masses,
  tol,
  n_isopeaks,
  min_isopeaks
)
}
\arguments{
\item{s}{\code{\link[MALDIquant]{MassSpectrum}} object}

\item{...}{Parameters passed to the different preprocessing functions.
See \code{\link{preprocess_data}}}
}
\value{
peaks list, with mass, intensity and s2n
}
\description{
preprocess_spectra
}
